add test and do second sweep of PR review
