+++ Cargo.toml
@@ -11,7 +11,6 @@ members = [
   "iroh-test",
   "iroh-net/bench",
   "iroh-cli",
-  "iroh-progress"
 ]
 resolver = "2"
