+++ iroh-net/Cargo.toml
@@ -130,7 +130,7 @@ default = ["metrics"]
 iroh-relay = ["clap", "toml", "rustls-pemfile", "regex", "serde_with", "tracing-subscriber"]
 metrics = ["iroh-metrics/metrics"]
 test-utils = ["axum"]
-local_node_discovery = ["swarm-discovery"]
+local_swarm_discovery = ["swarm-discovery"]
 
 [[bin]]
 name = "iroh-relay"
