+++ iroh-net/src/discovery.rs
@@ -12,10 +12,9 @@ use crate::{AddrInfo, Endpoint, NodeId};
 
 pub mod dns;
 
-/// enable local node discovery
-// TODO(ramfox): rename
-#[cfg(feature = "local_node_discovery")]
-pub mod local_node_discovery;
+/// enable local swarm discovery
+#[cfg(feature = "local_swarm_discovery")]
+pub mod local_swarm_discovery;
 pub mod pkarr;
 
 /// Name used for logging when new node addresses are added from discovery.
